Convert to relative or CDN URLs
In order to make it easier to do local development on patter, I have
changes all the paths to be relative. This will make it easier to
develop locallay, and also when deploying because paths won't need to
be differnet.

Also place where we load up plain jQuery I just used the google ajax
CDN. Not have to pay for assets is a good thing.